implement JsonOptions for converting dates
Before 2.3.1 dates where converted to UTC when ConvertToJson was called. However, ParseJson did not convert the times back to local times. This adds an option where the user can specify if dates should be converted (both ways) or not at all. As the default was that dates where converted, New variable option starts with"Dont", so the default would be set to false and closest to previous versions.